action: Move packages to internal/
